backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
# passlib 1.7.x expects bcrypt<4.1 (__about__ removed in bcrypt 4.1+)
bcrypt>=4.0.1,<4.1
pydantic
pydantic-settings
python-multipart
redis
alembic
pytest
httpx
